Copy file instead of move

+++ src/toil_scripts/gatk_germline/germline.py
@@ -142,7 +142,8 @@ def move_to_output_dir(work_dir, output_dir, *filenames):
         origin = os.path.join(work_dir, filename)
         dest = os.path.join(output_dir, filename)
         # With caching we cannot move the file
-        shutil.copy(origin, dest)
+	if not os.path.exists(dest):
+	    shutil.copy(origin, dest)
 
 
 def batch_start(job, input_args):
